requirements.txt
Django
djangorestframework
psycopg2-binary
django-bootstrap-v5
django-address